update2.0

+++ src/emulate_files/utilities.h
@@ -15,7 +15,7 @@ uint64_t readRegister (uint8_t registerIndex, uint8_t sf, uint64_t *generalRegis
 
 uint32_t extractBits(uint64_t n, uint8_t startIndex, uint8_t endIndex);
 
-uint8_t bitShift(uint8_t shift, uint64_t n, uint8_t operand);
+uint64_t bitShift(uint8_t shift, uint64_t n, uint8_t operand);
 
 void update_pstate(uint64_t result, uint64_t operand1, uint64_t operand2, bool is_subtraction, Pstate *pstate);
